refactor: 패키지 구조 재변경
